app/requirements.txt
flask
flask_api
uWSGI
pandas
pymysql
scikit-surprise
joblib
geopy